Modified audio_stream_player
Changed the error added to the set_volume_db method to a warning.

Now when the user sets an undefined value for the volume, they are notified through a warning.

The conditional created within the method was also changed so that if the value is undefined, the warning will trigger and the method will immediately end by returning. If the condition is not met, then the volume is changed to the one provided in the parameter.