Final Implementation of Solution
Created final implementation of audio bug solution.

Modified error check conditionals within the audio_stream_player file to address the bug. 

Now when a Godot script tries to insert an incorrect decibel value, a warning is prompted and the change is not performed. If a script tries to set an undefined decibel value, an error is thrown. If an audio player with an invalid volume value tries to play their sound, an error is thrown. 

Also included a Godot project file as a demonstration of our solution.